# serverStub/go/model_doctor.go
/*
 * Eyada API
 *
 * CRUD operations for Doctors and Patients, and the Eyade-Scheduler
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package eyadaAPI

type Doctor struct {
	// ID is generated before POST request. This is the public-facing id, i.e. shareable.
	Id int64 `json:"id"`

	Email string `json:"email"`

	FirstName string `json:"firstName,omitempty"`

	LastName string `json:"lastName,omitempty"`
	// Phone number must include country code, no non-numerical characters
	Phone int64 `json:"phone"`

	OrganizationIDs []string `json:"organizationIDs,omitempty"`
}