# serverStub/go/model_organization.go
/*
 * Eyada API
 *
 * CRUD operations for Doctors and Patients, and the Eyade-Scheduler
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package eyadaAPI

type Organization struct {
	// ID is generated before POST request. This is the public-facing id, i.e. shareable.
	Id string `json:"id"`

	Address *Address `json:"address,omitempty"`

	Email string `json:"email"`

	Number int64 `json:"number"`

	Doctors []string `json:"doctors,omitempty"`
}