# serverStub/go/model_patient.go
/*
 * Eyada API
 *
 * CRUD operations for Doctors and Patients, and the Eyade-Scheduler
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package eyadaAPI

type Patient struct {

	Id int64 `json:"id,omitempty"`

	Username string `json:"username,omitempty"`

	Address []Address `json:"address,omitempty"`
}